+++ packages/thumbprint-react/components/Image/components/image/index.jsx
