+++ packages/thumbprint-react/components/Image/index.jsx
@@ -1,2 +1,2 @@
-export { default as SmartImage } from './components/smart-image/index.jsx';
+export { default as Image } from './components/image/index.jsx';
 export { default as ImageServiceProvider } from './components/image-service-provider/index.jsx';
