+++ packages/thumbprint-react/index.js
@@ -7,7 +7,7 @@ export { default as Checkbox } from './components/Checkbox/index.jsx';
 export { default as DatePicker } from './components/DatePicker/index.jsx';
 export { default as FormNote } from './components/FormNote/index.jsx';
 export { Grid, GridColumn } from './components/Grid/index.jsx';
-export { SmartImage, ImageServiceProvider } from './components/Image/index.jsx';
+export { Image, ImageServiceProvider } from './components/Image/index.jsx';
 export { default as Input, InputIcon, InputClearButton } from './components/Input/index.jsx';
 export { default as InputRow, InputRowContext } from './components/InputRow/index.jsx';
 export { default as Label } from './components/Label/index.jsx';
